refactor(cached_object_store): continue seperate the files inside a module (#346)
this pr is a continue work on seperating the files in
cached_object_store after #295. all changes are code movements with no
logic modifications.

- moved LocalCacheStorage into `storage.rs`, while renamed
`fs_cache_storage.rs` to `storage_fs.rs`
- moved CachedObjectStore into `cached_object_store/object_store.rs`
- exposed the structs & traits in `mod.rs`
