+++ src/cached_object_store/mod.rs
@@ -1,0 +1,8 @@
+mod object_store;
+mod storage;
+mod storage_fs;
+
+pub(crate) use object_store::CachedObjectStore;
+#[allow(unused_imports)]
+pub use storage::{LocalCacheEntry, LocalCacheHead, LocalCacheStorage, PartID};
+pub use storage_fs::FsCacheStorage;
