+++ src/cached_object_store/object_store.rs
@@ -1,16 +1,13 @@
-use std::{collections::HashMap, fmt::Display, ops::Range, sync::Arc};
-
 use bytes::{Bytes, BytesMut};
 use futures::{future::BoxFuture, stream, stream::BoxStream, StreamExt};
 use object_store::{path::Path, GetOptions, GetResult, ObjectMeta, ObjectStore};
-use object_store::{Attribute, Attributes, GetRange, GetResultPayload, PutResult};
+use object_store::{Attributes, GetRange, GetResultPayload, PutResult};
 use object_store::{ListResult, MultipartUpload, PutMultipartOpts, PutOptions, PutPayload};
-use serde::{Deserialize, Serialize};
+use std::{ops::Range, sync::Arc};
 
+use crate::cached_object_store::storage::{LocalCacheStorage, PartID};
 use crate::error::SlateDBError;
 use crate::metrics::DbStats;
-
-pub(crate) mod fs_cache_storage;
 
 #[derive(Debug, Clone)]
 pub(crate) struct CachedObjectStore {
@@ -423,73 +420,6 @@ impl ObjectStore for CachedObjectStore {
     }
 }
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-struct LocalCacheHead {
-    pub location: String,
-    pub last_modified: String,
-    pub size: usize,
-    pub e_tag: Option<String>,
-    pub version: Option<String>,
-    pub attributes: HashMap<String, String>,
-}
-
-impl LocalCacheHead {
-    pub fn meta(&self) -> ObjectMeta {
-        ObjectMeta {
-            location: self.location.clone().into(),
-            last_modified: self.last_modified.parse().unwrap_or_default(),
-            size: self.size,
-            e_tag: self.e_tag.clone(),
-            version: self.version.clone(),
-        }
-    }
-
-    pub fn attributes(&self) -> Attributes {
-        let mut attrs = Attributes::new();
-        for (key, value) in self.attributes.iter() {
-            let key = match key.as_str() {
-                "Cache-Control" => Attribute::CacheControl,
-                "Content-Disposition" => Attribute::ContentDisposition,
-                "Content-Encoding" => Attribute::ContentEncoding,
-                "Content-Language" => Attribute::ContentLanguage,
-                "Content-Type" => Attribute::ContentType,
-                _ => Attribute::Metadata(key.to_string().into()),
-            };
-            let value = value.to_string().into();
-            attrs.insert(key, value);
-        }
-        attrs
-    }
-}
-
-impl From<(&ObjectMeta, &Attributes)> for LocalCacheHead {
-    fn from((meta, attrs): (&ObjectMeta, &Attributes)) -> Self {
-        let mut attrs_map = HashMap::new();
-        for (key, value) in attrs.iter() {
-            let key = match key {
-                Attribute::CacheControl => "Cache-Control",
-                Attribute::ContentDisposition => "Content-Disposition",
-                Attribute::ContentEncoding => "Content-Encoding",
-                Attribute::ContentLanguage => "Content-Language",
-                Attribute::ContentType => "Content-Type",
-                Attribute::Metadata(key) => key,
-                _ => continue,
-            };
-            attrs_map.insert(key.to_string(), value.to_string());
-        }
-        LocalCacheHead {
-            location: meta.location.to_string(),
-            last_modified: meta.last_modified.to_rfc3339(),
-            size: meta.size,
-            e_tag: meta.e_tag.clone(),
-            version: meta.version.clone(),
-            attributes: attrs_map,
-        }
-    }
-}
-
-// it seems that object_store did not expose this error type, duplicate it here.
-// TODO: raise a pr to expose this error type in object_store.
 #[derive(Debug, thiserror::Error)]
 pub(crate) enum InvalidGetRange {
     #[error("Range start too large, requested: {requested}, length: {length}")]
@@ -499,58 +429,18 @@ pub(crate) enum InvalidGetRange {
     Inconsistent { start: usize, end: usize },
 }
 
-#[async_trait::async_trait]
-pub trait LocalCacheStorage: Send + Sync + std::fmt::Debug + Display + 'static {
-    fn entry(
-        &self,
-        location: &object_store::path::Path,
-        part_size: usize,
-    ) -> Box<dyn LocalCacheEntry>;
-
-    async fn start_evictor(&self);
-}
-
-#[async_trait::async_trait]
-pub trait LocalCacheEntry: Send + Sync + std::fmt::Debug + 'static {
-    async fn save_part(&self, part_number: PartID, buf: Bytes) -> object_store::Result<()>;
-
-    async fn read_part(
-        &self,
-        part_number: PartID,
-        range_in_part: Range<usize>,
-    ) -> object_store::Result<Option<Bytes>>;
-
-    /// might be useful on rewriting GET request on the prefetch phase. the cached files are
-    /// expected to be in the same folder, so it'd be expected to be fast without expensive
-    /// globbing.
-    #[cfg(test)]
-    async fn cached_parts(&self) -> object_store::Result<Vec<PartID>>;
-
-    async fn save_head(&self, meta: (&ObjectMeta, &Attributes)) -> object_store::Result<()>;
-
-    async fn read_head(&self) -> object_store::Result<Option<(ObjectMeta, Attributes)>>;
-}
-
-pub(crate) type PartID = usize;
-
 #[cfg(test)]
 mod tests {
     use std::sync::Arc;
 
-    use bytes::Bytes;
     use object_store::{path::Path, GetOptions, GetRange, ObjectStore, PutPayload};
-    use rand::{thread_rng, Rng};
+    use rand::Rng;
 
     use super::CachedObjectStore;
-    use crate::cached_object_store::fs_cache_storage::FsCacheStorage;
-    use crate::cached_object_store::{fs_cache_storage::FsCacheEntry, PartID};
+    use crate::cached_object_store::storage_fs::FsCacheStorage;
+    use crate::cached_object_store::{storage::PartID, storage_fs::FsCacheEntry};
     use crate::metrics::DbStats;
-
-    pub(crate) fn gen_rand_bytes(n: usize) -> Bytes {
-        let mut rng = thread_rng();
-        let random_bytes: Vec<u8> = (0..n).map(|_| rng.gen()).collect();
-        Bytes::from(random_bytes)
-    }
+    use crate::test_utils::gen_rand_bytes;
 
     fn new_test_cache_folder() -> std::path::PathBuf {
         let mut rng = rand::thread_rng();
